+++ src/main/java/org/zalando/nakadi/controller/SettingsController.java
@@ -17,6 +17,7 @@ import org.zalando.nakadi.domain.ItemsWrapper;
 import org.zalando.nakadi.domain.ResourceAuthorization;
 import org.zalando.nakadi.exceptions.UnableProcessException;
 import org.zalando.nakadi.exceptions.runtime.UnknownOperationException;
+import org.zalando.nakadi.plugin.api.authz.AuthorizationService;
 import org.zalando.nakadi.security.Client;
 import org.zalando.nakadi.service.AdminService;
 import org.zalando.nakadi.service.BlacklistService;
@@ -28,16 +29,17 @@ import org.zalando.problem.spring.web.advice.Responses;
 import javax.validation.Valid;
 import javax.ws.rs.core.Response;
 
+import static org.zalando.nakadi.domain.AdminResource.ADMIN_RESOURCE;
+
 @RestController
 @RequestMapping(value = "/settings")
 public class SettingsController {
 
+    private static final Logger LOG = LoggerFactory.getLogger(SettingsController.class);
     private final BlacklistService blacklistService;
     private final FeatureToggleService featureToggleService;
     private final SecuritySettings securitySettings;
     private final AdminService adminService;
-
-    private static final Logger LOG = LoggerFactory.getLogger(SettingsController.class);
 
 
     @Autowired
@@ -52,8 +54,8 @@ public class SettingsController {
     }
 
     @RequestMapping(path = "/blacklist", method = RequestMethod.GET)
-    public ResponseEntity<?> getBlacklist(final Client client) {
-        if (isNotAdmin(client)) {
+    public ResponseEntity<?> getBlacklist() {
+        if (!adminService.isAdmin(AuthorizationService.Operation.READ)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         return ResponseEntity.ok(blacklistService.getBlacklist());
@@ -61,9 +63,8 @@ public class SettingsController {
 
     @RequestMapping(value = "/blacklist/{blacklist_type}/{name}", method = RequestMethod.PUT)
     public ResponseEntity blacklist(@PathVariable("blacklist_type") final BlacklistService.Type blacklistType,
-                                        @PathVariable("name") final String name,
-                                        final Client client) {
-        if (isNotAdmin(client)) {
+                                    @PathVariable("name") final String name) {
+        if (!adminService.isAdmin(AuthorizationService.Operation.WRITE)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         blacklistService.blacklist(name, blacklistType);
@@ -72,9 +73,8 @@ public class SettingsController {
 
     @RequestMapping(value = "/blacklist/{blacklist_type}/{name}", method = RequestMethod.DELETE)
     public ResponseEntity whitelist(@PathVariable("blacklist_type") final BlacklistService.Type blacklistType,
-                                         @PathVariable("name") final String name,
-                                         final Client client) {
-        if (isNotAdmin(client)) {
+                                    @PathVariable("name") final String name) {
+        if (!adminService.isAdmin(AuthorizationService.Operation.WRITE)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         blacklistService.whitelist(name, blacklistType);
@@ -83,7 +83,7 @@ public class SettingsController {
 
     @RequestMapping(path = "/features", method = RequestMethod.GET)
     public ResponseEntity<?> getFeatures(final Client client) {
-        if (isNotAdmin(client)) {
+        if (!adminService.isAdmin(AuthorizationService.Operation.READ)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         return ResponseEntity.ok(new ItemsWrapper<>(featureToggleService.getFeatures()));
@@ -92,7 +92,7 @@ public class SettingsController {
     @RequestMapping(path = "/features", method = RequestMethod.POST)
     public ResponseEntity<?> setFeature(@RequestBody final FeatureToggleService.FeatureWrapper featureWrapper,
                                         final Client client) {
-        if (isNotAdmin(client)) {
+        if (!adminService.isAdmin(AuthorizationService.Operation.WRITE)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
         featureToggleService.setFeature(featureWrapper);
@@ -101,12 +101,18 @@ public class SettingsController {
 
     @RequestMapping(path = "/admins", method = RequestMethod.GET)
     public ResponseEntity<?> getAdmins() {
+        if (!adminService.isAdmin(AuthorizationService.Operation.READ)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(ResourceAuthorization.fromPermissionsList(adminService.getAdmins()));
     }
 
     @RequestMapping(path = "/admins", method = RequestMethod.POST)
     public ResponseEntity<?> updateAdmins(@Valid @RequestBody final ResourceAuthorization authz) {
-        adminService.updateAdmins(authz.toPermissionsList("nakadi"));
+        if (!adminService.isAdmin(AuthorizationService.Operation.ADMIN)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        adminService.updateAdmins(authz.toPermissionsList(ADMIN_RESOURCE));
         return ResponseEntity.ok().build();
     }