+++ src/test/java/org/zalando/nakadi/controller/CursorsControllerTest.java
@@ -112,7 +112,7 @@ public class CursorsControllerTest {
 
     @Test
     public void whenCommitValidCursorsThenNoContent() throws Exception {
-        when(cursorsService.commitCursors(any(), any(), any(), any()))
+        when(cursorsService.commitCursors(any(), any(), any()))
                 .thenReturn(ImmutableList.of());
         postCursors(DUMMY_CURSORS)
                 .andExpect(status().isNoContent());
@@ -120,7 +120,7 @@ public class CursorsControllerTest {
 
     @Test
     public void whenCommitInvalidCursorsThenOk() throws Exception {
-        when(cursorsService.commitCursors(any(), any(), any(), any()))
+        when(cursorsService.commitCursors(any(), any(), any()))
                 .thenReturn(DUMMY_CURSORS.stream().map(v -> Boolean.FALSE).collect(Collectors.toList()));
         final ItemsWrapper<CursorCommitResult> expectation = new ItemsWrapper<>(
                 DUMMY_CURSORS.stream()
@@ -133,7 +133,7 @@ public class CursorsControllerTest {
 
     @Test
     public void whenNoSubscriptionThenNotFound() throws Exception {
-        when(cursorsService.commitCursors(any(), eq(SUBSCRIPTION_ID), any(), any()))
+        when(cursorsService.commitCursors(any(), eq(SUBSCRIPTION_ID), any()))
                 .thenThrow(new NoSuchSubscriptionException("dummy-message"));
         final Problem expectedProblem = Problem.valueOf(NOT_FOUND, "dummy-message");
 
@@ -142,7 +142,7 @@ public class CursorsControllerTest {
 
     @Test
     public void whenNoEventTypeThenUnprocessableEntity() throws Exception {
-        when(cursorsService.commitCursors(any(), any(), any(), any()))
+        when(cursorsService.commitCursors(any(), any(), any()))
                 .thenThrow(new NoSuchEventTypeException("dummy-message"));
         final Problem expectedProblem = Problem.valueOf(UNPROCESSABLE_ENTITY, "dummy-message");
 
@@ -151,7 +151,7 @@ public class CursorsControllerTest {
 
     @Test
     public void whenServiceUnavailableExceptionThenServiceUnavailable() throws Exception {
-        when(cursorsService.commitCursors(any(), any(), any(), any()))
+        when(cursorsService.commitCursors(any(), any(), any()))
                 .thenThrow(new ServiceUnavailableException("dummy-message"));
         final Problem expectedProblem = Problem.valueOf(SERVICE_UNAVAILABLE, "dummy-message");
 
@@ -160,7 +160,7 @@ public class CursorsControllerTest {
 
     @Test
     public void whenInvalidCursorExceptionThenUnprocessableEntity() throws Exception {
-        when(cursorsService.commitCursors(any(), any(), any(), any()))
+        when(cursorsService.commitCursors(any(), any(), any()))
                 .thenThrow((new InvalidCursorException(CursorError.NULL_PARTITION,
                         new SubscriptionCursor(null, null, null, null))));