+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - New database table for authorization
 - Endpoints for getting and updating the lists of administrators
 - Default administrator defined in application.yml
+- Allow admins to bypass event type authorization restrictions
+
+### Changed
+- Admin endpoints use list of admins from Postgres
 
 ## [2.2.0] - 2017-08-29